# [Enhance] Add prepare phase for some timestamp functions