--- a/be/src/common/daemon.cpp
+++ b/be/src/common/daemon.cpp
@@ -53,7 +53,6 @@
 #include "exprs/json_functions.h"
 #include "exprs/hll_hash_function.h"
 #include "exprs/grouping_sets_functions.h"
-#include "exprs/timezone_db.h"
 #include "exprs/bitmap_function.h"
 #include "exprs/hll_function.h"
 #include "geo/geo_functions.h"
@@ -275,7 +274,6 @@ void init_daemon(int argc, char** argv, const std::vector<StorePath>& paths) {
     ESFunctions::init();
     GeoFunctions::init();
     GroupingSetsFunctions::init();
-    TimezoneDatabase::init();
     BitmapFunctions::init();
     HllFunctions::init();
     HashFunctions::init();