--- a/be/src/runtime/datetime_value.h
+++ b/be/src/runtime/datetime_value.h
@@ -462,34 +462,6 @@ public:
 
     void set_type(int type);
 
-    static bool find_cctz_time_zone(const std::string& timezone, cctz::time_zone& ctz) {
-        re2::StringPiece value;
-        if (time_zone_offset_format_reg.Match(timezone, 0, timezone.size(), RE2::UNANCHORED, &value, 1)) {
-            bool postive = value[0] != '-';
-
-            //Regular expression guarantees hour and minute mush be int 
-            int hour = std::stoi(value.substr(1, 2).as_string());
-            int minute = std::stoi(value.substr(4, 2).as_string());
-
-            // timezone offsets around the world extended from -12:00 to +14:00
-            if (!postive && hour > 12) {
-                return false;
-            } else if (postive && hour > 14) {
-                return false;
-            }
-            int offset = hour * 60 * 60 + minute * 60;
-            offset *= postive ? 1 : -1;
-            ctz = cctz::fixed_time_zone(cctz::seconds(offset));
-            return true;
-        } else if (timezone == "CST"){
-            // Supports offset and region timezone type, "CST" use here is compatibility purposes.
-            ctz = cctz::fixed_time_zone(cctz::seconds(8 * 60 * 60));
-            return true;
-        } else {
-            return cctz::load_time_zone(timezone, &ctz);
-        }
-    }
-
 private:
     // Used to make sure sizeof DateTimeValue
     friend class UnusedClass;