… be/src/exprs/CMakeLists.txt
… be/src/exprs/timezone_db.cpp
… be/src/runtime/runtime_state.h
… be/src/udf/udf.cpp
… be/src/util/CMakeLists.txt
… be/src/util/timezone_utils.cpp
… be/src/util/timezone_utils.h
… build.sh
… docs/en/sql-reference/sql-functions/date-time-functions/to_date.md
… docs/zh-CN/sql-reference/sql-functions/date-time-functions/convert_tz.md
… docs/zh-CN/sql-reference/sql-functions/date-time-functions/to_date.md
… fe/src/main/java/org/apache/doris/rewrite/FEFunctions.java
… gensrc/script/doris_builtins_functions.py
… run-ut.sh